Add initial draft of priority boosting by build duration
This adds a new utility methods which is designed to boost the priority
of an item based on its build duration.  Builds which are considered
fast as defined by a provided threshold have their priorities boosted
higher.  Builds that are considered slow have their priorities reduced.

The boosting takes into account how long a build has been waiting so
the effects of boosting may may be eliminated/reversed depending on
the wait time of a build relative to its duration.